Better control over ripple
Current implementation does not provide enough control over Ripple effect.

In particular, the ripple effect should not be based on the dropdown opacity animation duration but rather have its own duration.
That's also the case for the animation delay.

For example, if I want to disable totally the Ripple effect, I currently have no choice than setting opacity 0 to the ripple effect. But this is not enough, because of the delay, the first feedback the user see is delayed. And I don't want to set the opacity animation to 0 to remove that delay.

This PR permit to easily disable the ripple effect, and decouple ripple animation duration from dropdown opacity animation duration